# memorybank/docs/api/openapi-3.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DonorDesk Public API",
    "description": "REST API for DonorDesk humanitarian aid reporting platform. Provides programmatic access to projects, evidence, reports, and compliance data for partner integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "DonorDesk Support",
      "email": "api@donordesk.org"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://donordesk.org/legal"
    }
  },
  "servers": [
    {
      "url": "https://api.donordesk.org/v1",
      "description": "Production"
    },
    {
      "url": "https://staging-api.donordesk.org/v1",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/projects/{projectId}/evidence": {
      "get": {
        "operationId": "listEvidence",
        "summary": "List evidence for a project",
        "tags": ["Evidence"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/PageToken"
          }
        ],
        "responses": {
          "200": {
            "description": "List of evidence items",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceList"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "uploadEvidence",
        "summary": "Upload evidence",
        "description": "Upload a new evidence file to a project",
        "tags": ["Evidence"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evidence uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /auth/token"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key for webhook receivers and integrations"
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": { "type": "integer", "default": 20, "maximum": 100 }
      },
      "PageToken": {
        "name": "pageToken",
        "in": "query",
        "schema": { "type": "string" }
      }
    },
    "schemas": {
...
      "Evidence": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "fileName": { "type": "string" },
          "fileType": { "type": "string" },
          "uploadedAt": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["DRAFT", "SUBMITTED", "VERIFIED"] },
          "aiSummary": { "type": "string" },
          "extractedText": { "type": "string", "description": "Raw document text extracted by the Tika evidence_parse flow; consumed by AI report generation (2026-08-17)" }
        }
      },
      "EvidenceList": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Evidence" } },
          "nextPageToken": { "type": "string" }
        }
      },
      "EvidenceUploadRequest": {
        "type": "object",
        "required": ["file", "title"],
        "properties": {
          "file": { "type": "string", "format": "binary" },
          "title": { "type": "string" },
          "description": { "type": "string" }
        }
      },
...
    },
...
  },
...
}